// src/prisma/error-code.enum.ts
export enum PrismaErrorCode {
  VALUE_TOO_LONG = 'P2000',
  RECORD_NOT_FOUND = 'P2001',
  UNIQUE_CONSTRAINT_FAILED = 'P2002',
  FOREIGN_KEY_CONSTRAINT_FAILED = 'P2003',
  CONSTRAINT_FAILED = 'P2004',
  INVALID_VALUE_STORED = 'P2005',
  INVALID_VALUE_PROVIDED = 'P2006',
  DATA_VALIDATION_ERROR = 'P2007',
  QUERY_PARSE_FAILED = 'P2008',
  QUERY_VALIDATION_FAILED = 'P2009',
  RAW_QUERY_FAILED = 'P2010',
  NULL_CONSTRAINT_VIOLATION = 'P2011',
  MISSING_REQUIRED_VALUE = 'P2012',
  MISSING_REQUIRED_ARGUMENT = 'P2013',
  REQUIRED_RELATION_VIOLATION = 'P2014',
  RELATED_RECORD_NOT_FOUND = 'P2015',
  QUERY_INTERPRETATION_ERROR = 'P2016',
  RECORDS_NOT_CONNECTED = 'P2017',
  REQUIRED_CONNECTED_RECORDS_NOT_FOUND = 'P2018',
  INPUT_ERROR = 'P2019',
  VALUE_OUT_OF_RANGE = 'P2020',
  TABLE_NOT_FOUND = 'P2021',
  COLUMN_NOT_FOUND = 'P2022',
  INCONSISTENT_COLUMN_DATA = 'P2023',
  CONNECTION_POOL_TIMEOUT = 'P2024',
  RECORD_REQUIRED_BUT_NOT_FOUND = 'P2025',
  FEATURE_NOT_SUPPORTED = 'P2026',
  MULTIPLE_ERRORS_OCCURRED = 'P2027',
  TRANSACTION_API_ERROR = 'P2028',
  QUERY_PARAMETER_LIMIT_EXCEEDED = 'P2029',
  FULLTEXT_INDEX_NOT_FOUND = 'P2030',
  NUMBER_OUT_OF_RANGE_FOR_INT64 = 'P2033',
  TRANSACTION_DEADLOCK = 'P2034',
  ASSERTION_VIOLATION = 'P2035',
  EXTERNAL_CONNECTOR_ERROR = 'P2036',
  TOO_MANY_DATABASE_CONNECTIONS = 'P2037',
}